added the dimensions and location forms components to the publish space area
