+++ src/index.css
@@ -1,9 +1,14 @@
 body {
   margin: 0;
-  display: flex;
+  padding: 0;
 }
 
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+*,*::after,*::before{
+  box-sizing: border-box;
+}
+
